pyproject.toml
[tool.poetry]
name = "projects-and-contracts"
version = "0.1.0"
description = ""
authors = ["Aliaksei Kastsiuchonak <kostkost30@gmail.com>"]
readme = "README.md"
packages = [{include = "projects_and_contracts"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "2.9.7"
python-dotenv = "1.0.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.4.0"

[tool.mypy]
mypy_path = "/"
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
no_implicit_optional = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

[tool.black]
skip-string-normalization = false
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
